fuzz: Move-only net utils
