+++ src/test/fuzz/util/mempool.h
@@ -5,10 +5,12 @@
 #ifndef BITCOIN_TEST_FUZZ_UTIL_MEMPOOL_H
 #define BITCOIN_TEST_FUZZ_UTIL_MEMPOOL_H
 
-#include <primitives/transaction.h>
-#include <test/fuzz/FuzzedDataProvider.h>
-#include <txmempool.h>
+#include <txmempool_entry.h>
 #include <validation.h>
+
+class CTransaction;
+class CTxMemPool;
+class FuzzedDataProvider;
 
 class DummyChainState final : public Chainstate
 {
